src/backend/requirements.txt
beautifulsoup4~=4.9.3
paramiko~=2.7.2
Pillow~=8.0.1
python-dateutil~=2.8.1
requests~=2.25.1
scrapy~=2.4.1
SQLAlchemy~=1.3.22
w3lib~=1.22.0
crochet~=1.12.0
Flask~=1.1.2
PyYAML~=5.3.1
pandas~=1.2.0
numpy~=1.19.4
wheel